Moved development library injection to template post-hooks and performed code cleanup #207
* Moved bootstrap code from CLI to utils.py
* Moved bootstrap calls to post-hooks in templates
* Moved convert_line_endings to utils.py
* Changed dev distribution directory name to devdist
