+++ evalutils/templates/evaluator/{{ cookiecutter.package_name }}/Dockerfile
@@ -17,8 +17,9 @@ ENV PATH="/home/evaluator/.local/bin:${PATH}"
 RUN python -m pip install --user -U pip
 
 {% if cookiecutter.dev_build|int -%}
-COPY --chown=evaluator:evaluator evalutils /opt/evaluation/evalutils
-RUN python -m pip install --user /opt/evaluation/evalutils
+# Installs development distributions of evalutils
+COPY --chown=evaluator:evaluator devdist /opt/evaluation/devdist
+RUN python -m pip install --user /opt/evaluation/devdist
 {%- endif %}
 
 COPY --chown=evaluator:evaluator ground-truth /opt/evaluation/ground-truth
