+++ evalutils/templates/processor/hooks/post_gen_project.py
@@ -2,38 +2,25 @@
 import os
 import shutil
 from pathlib import Path
+from evalutils.utils import (
+    bootstrap_development_distribution,
+    convert_line_endings,
+)
 
-EOL_UNIX = b"\n"
-EOL_WIN = b"\r\n"
-EOL_MAC = b"\r"
 
-templated_python_files = Path(os.getcwd()).glob("*.py.j2")
+PROCESSOR_NAME = "{{ cookiecutter.processor_name }}"
+IS_DEV_BUILD = {{cookiecutter.dev_build}} == 1
+
+template_dir = Path(os.getcwd())
+
+templated_python_files = template_dir.glob("*.py.j2")
 for f in templated_python_files:
     shutil.move(f.name, f.stem)
 
 
-def convert_line_endings():
-    """ Enforce unix line endings for the generated files """
-    files = []
-    for ext in [
-        ".py",
-        ".sh",
-        "Dockerfile",
-        ".txt",
-        ".csv",
-        ".mhd",
-        ".gitignore",
-    ]:
-        files.extend(Path(".").glob(f"**/*{ext}"))
-
-    for file in files:
-        with open(file, "rb") as f:
-            lines = f.read()
-
-        lines = lines.replace(EOL_WIN, EOL_UNIX).replace(EOL_MAC, EOL_UNIX)
-
-        with open(file, "wb") as f:
-            f.write(lines)
-
-
 convert_line_endings()
+
+if IS_DEV_BUILD:
+    bootstrap_development_distribution(
+        PROCESSOR_NAME, template_dir / "devdist"
+    )
